power: safely convert int to void *

+++ services/core/jni/com_android_server_power_PowerManagerService.cpp
@@ -167,14 +167,14 @@ static void nativeSetFeature(JNIEnv *env, jclass clazz, jint featureId, jint dat
 static void nativeCpuBoost(JNIEnv *env, jobject clazz, jint duration) {
     // Tell the Power HAL to boost the CPU
     if (gPowerModule && gPowerModule->powerHint) {
-        gPowerModule->powerHint(gPowerModule, POWER_HINT_CPU_BOOST, (void *) duration);
+        gPowerModule->powerHint(gPowerModule, POWER_HINT_CPU_BOOST, (void *)(intptr_t) duration);
     }
 }
 
 static void nativeSetPowerProfile(JNIEnv *env, jobject clazz, jint profile) {
     // Tell the Power HAL to select a power profile
     if (gPowerModule && gPowerModule->powerHint) {
-        gPowerModule->powerHint(gPowerModule, POWER_HINT_SET_PROFILE, (void *) profile);
+        gPowerModule->powerHint(gPowerModule, POWER_HINT_SET_PROFILE, (void *)(intptr_t) profile);
     }
 }
